updated for class structure
