--- mrpir-service.service
@@ -1,30 +1,0 @@
-[Unit]
-Description=Presense sensor service
-# After=multi-user.target
-# Wants=network-online.target systemd-networkd-wait-online.service
-
-# StartLimitIntervalSec=0
-# StartLimitBurst=0
-
-[Service]
-# Our service will notify systemd once it is up and running
-Type=notify
-Restart=on-failure
-# RestartSec=5s
-
-# WorkingDirectory=/home/michael/mrpir
-# RootDirectory=/usr/bin
-
-# Disable Python's buffering of STDOUT and STDERR, so that output from the
-# service shows up immediately in systemd's logs
-Environment=PYTHONUNBUFFERED=1
-
-# Command to execute when the service is started
-ExecStart=/usr/bin/python3 /home/michael/mrpir/testservice.py
-
-# User=michael
-# Group=users
-
-[Install]
-# WantedBy=multi-user.target
-WantedBy=default.target
