--- mrtest.service
@@ -1,20 +1,0 @@
-[Unit]
-Description=Python Demo Service
-
-[Service]
-Environment=PYTHONUNBUFFERED=1
-Type=notify
-Restart=on-failure
-RestartSec=5s
-StartLimitInterval=60s
-StartLimitBurst=3
-TimeoutSec=15
-WatchdogSec=10
-
-
-WorkingDirectory=/home/michael/mrpir
-# Environment="PYTHONPATH=$PYTHONPATH:/home/michael/.local/lib/python3.6/site-packages"
-# Command to execute when the service is started
-ExecStart=python3 /home/michael/mrpir/testservice.py --serve-in-foreground
-
-# user=michael
